MLK-19589-01 ARM: imx: add imx6ulz basic support
The i.MX6ULZ is new SOC of the i.MX6 series. it
is SW compatile with i.MX6ULL, so most of the code
can be reuse from i.MX6ULL. To maximum the SW reuse,
i.MX6ULZ don't have an independent SOC id in anamix.
so a dummy ID is used to identify it.

Signed-off-by: Bai Ping <ping.bai@nxp.com>
Signed-off-by: Eran Matityahu <eran.m@variscite.com>