imx93-var-som-jig: dts: use a reserved mem-ranges to constrain ele-mu-dma-range
sentinel cannot read/write DDR in the range
- 0xE0000000  0xFFFFFFFF

while genrating random numbers with hwrng device,
sometimes output address is allocated from the range
which is not accessible to sentinel.

so we are reserving 1MB of ddr memory
to make sure that the memory is allocated
from the reserved mem range.

Patch derived from:
https://github.com/nxp-imx/linux-imx/commit/1cef27794d6d

Signed-off-by: Ken Sloat <ken.s@variscite.com>